Massive (bad) rearchitecture to avoid some level of code duplication
Deployment scripts not included - use your own. But as a first step, make sure to copy the contenst of shared.
Yes, mono repo bad. I should make my own convenience NPM packages, but that'll be a next step.
The collector stores segments in the DB, and runs (right now) at an hour cadence. (you are free to do this as often as you wish, however)
The reporter fetches segments, and runs nightly for a daily report, and a monthly report.
Couple of bad things I'll note:
1. bucket/db/mapAPI ideally should be as a lightweight NPM package. Will do that later to simplify build process.
2. I'm doing the deploys manually, and configuring the cron jobs manually. Ideally should do the cron job configuration programmatically.
3. Using the actual name as the lookup key in psql is bad, I know. Will want to assign a unique 'segment id' or 'trip id', but for now, this stays
4. Package.json isn't shared well, either. Extending this will be a nightmare.